requirements.txt
telethon==1.42.0
baidu-aip
opencv-python-headless
numpy
